Add text file input to run_make_histograms.C
Previously, run_make_histograms.C would have to be edited in order to
change the files or dates it passed to make_histograms, which is why
there were hitemp and lowtemp versions of it. It was clunky and very
difficult to modify its behavior. Now, a single macro named
run_make_histograms.C takes five text files as input: four containing
the filenames of the runs on different tracker partitions, and one
containing the (approximate - just has to be within the same month)
corresponding rundates.

In order to demonstrate the way this input ability would be used, I have
created run_make_histos.sh, which contains a command to run
run_make_histograms.C with all necessary inputs.