Make ref run numbers as inputs to draw.C
draw.C now takes the ref run numbers as inputs instead of having
hardcoded ref run dates. The run dates can always be determined by
looking up the run numbers; in this sense the run numbers are more
precise. Note that they are still passed as strings since they are used
in plotting commands as part of an axis label.

In addition I eliminated run_hitemp_draw.C and run_lowtemp_draw.C in
favor of calling the scripts as lines in shell scripts - run_4c_draw.sh
is an example of how this would be done for the 4 C runs.