Add y-axis parameter editing, fix all errors
Added an input to make_histograms.C which gives it a file containing, in
the order in which the histograms are computed, the number of bins, min,
and max of each histogram's y-axis. This was set up in order to make it
easier to edit the binning and size of the y-axis so that zooming in on
the relevant features in a given plot doesn't have to involve editing
the code itself.

Also ran the full plotting workflow (produce_4c_histos.sh) and fixed all
errors that came up. This commit should work for any set of runs.