Make diff independent of run files
Prior to this, differences in values of the variables were obtained by
reading in the "diff" branch of the run trees. This was a problem
because, in the situation that one decided to change the reference run
for a given partition, one would have to re-download all run files for
that partition from tkCommissioner, selecting the new reference run for
each one. Now make_histograms takes in a ref filename and subtracts the
ref value from the run value to get the diff. This way, one can keep a
set of run files (with no reference run specified in tkCommissioner) and
simply change the name of the ref file whenever a ref change is needed.

Changes are reflected in the filenames as they no longer contain a ref
run number, as can be seen in the filenames.txt files.