Fix matching criteria and method
Previously, the criterion for matching rows between run and ref trees
was the deviceID, which is not independent of the database partition in
which runs are stored and thus led to problems. Now the rows are matched
by DetID and I2CAddress, as the combination of those two does not vary
for a given device.

Additionally, the method by which the matching was done was arcane and
slow, using the Draw method on the ref tree to return entry numbers of
matching rows. Now the matching is done by building an index on the ref
tree and then using GetEntryWithIndex to get the corresponding row for
each one in the run tree, which is significantly faster.