Reconfigure hists to fill independently
Rewrote run_make_histograms.C into run_make_partition_hists.C, which has
better encapsulation of certain functions (such as parsing the config
files, which are now in their own directory) and fills the histograms by
partition. Previously the gainscan runs we used were clusters of runs on
all partitions, but now this restriction is not needed, allowing us to
fill in the gaps in the plots.