Create utility for graphing performance by APV
make_tgraphs.C creates individual TGraphs for each APV in a given
partition. It takes as input a config file with a list of filenames (the
format is the same as that for make_histograms to reduce effort) and an
output filename. It outputs a root file containing TGraphs named in the
scheme <DetID>_<I2CAddress>.